+++ lib/komponent/komponent_helper.rb
@@ -7,6 +7,7 @@ module KomponentHelper
     captured_block = proc { |args| capture(args, &block) } if block_given?
     Komponent::ComponentRenderer.new(
       controller,
+      view_flow || (view && view.view_flow),
     ).render(
       component_name,
       locals,