+++ fixtures/my_app/frontend/components/partial/index.js
@@ -1,0 +1,1 @@
+import "components/partial/universe/partial_universe";